Revert changes in OverseerSolrRespose and move serialization
