+++ solr/core/src/java/org/apache/solr/cloud/OverseerSolrResponse.java
@@ -17,17 +17,11 @@
 package org.apache.solr.cloud;
 
 import org.apache.solr.client.solrj.SolrResponse;
-import org.apache.solr.common.SolrException;
-import org.apache.solr.common.SolrException.ErrorCode;
 import org.apache.solr.common.util.NamedList;
-import org.apache.solr.common.util.Utils;
-
-import java.io.IOException;
-import java.util.Objects;
 
 public class OverseerSolrResponse extends SolrResponse {
   
-  NamedList<Object> responseList = null;
+  NamedList responseList = null;
 
   private long elapsedTime;
   
@@ -53,51 +47,6 @@ public class OverseerSolrResponse extends SolrResponse {
   @Override
   public NamedList<Object> getResponse() {
     return responseList;
-  }
-
-  /**
-   * This method serializes the content of an {@code OverseerSolrResponse}. Note that:
-   * <ul>
-   * <li>The elapsed time is not serialized</li>
-   * <li>"Unknown" elements for the Javabin format will be serialized as Strings. See {@link org.apache.solr.common.util.JavaBinCodec#writeVal}</li>
-   * </ul>
-   */
-  @SuppressWarnings("deprecation")
-  public static byte[] serialize(OverseerSolrResponse responseObject) {
-    Objects.requireNonNull(responseObject);
-    if (useUnsafeSerialization()) {
-      return SolrResponse.serializable(responseObject);
-    }
-    try {
-      return Utils.toJavabin(responseObject.getResponse()).readAllBytes();
-    } catch (IOException|RuntimeException e) {
-      throw new SolrException(ErrorCode.SERVER_ERROR, "Exception serializing response to Javabin", e);
-    }
-  }
-  
-  static boolean useUnsafeSerialization() {
-    String useUnsafeOverseerResponse = System.getProperty("solr.useUnsafeOverseerResponse");
-    return useUnsafeOverseerResponse != null && ("true".equals(useUnsafeOverseerResponse));
-  }
-  
-  static boolean useUnsafeDeserialization() {
-    String useUnsafeOverseerResponse = System.getProperty("solr.useUnsafeOverseerResponse");
-    return useUnsafeOverseerResponse != null && ("true".equals(useUnsafeOverseerResponse) || "deserialization".equals(useUnsafeOverseerResponse));
-  }
-
-  @SuppressWarnings("deprecation")
-  public static OverseerSolrResponse deserialize(byte[] responseBytes) {
-    Objects.requireNonNull(responseBytes);
-    try {
-      @SuppressWarnings("unchecked")
-      NamedList<Object> response = (NamedList<Object>) Utils.fromJavabin(responseBytes);
-      return new OverseerSolrResponse(response);
-    } catch (IOException|RuntimeException e) {
-      if (useUnsafeDeserialization()) {
-        return (OverseerSolrResponse) SolrResponse.deserialize(responseBytes);
-      }
-      throw new SolrException(ErrorCode.SERVER_ERROR, "Exception deserializing response from Javabin", e);
-    }
   }
   
 }
